Add support for building for both iOS and macOS platforms

+++ Meadow/Core/Terrain/TerrainLayer.swift
@@ -41,7 +41,7 @@ public class TerrainLayer {
         
         for index in 0..<corners.count {
             
-            vertices.append(SCNVector3(x: CGFloat(node.volume.coordinate.x) + unit.vertices[index].x, y: World.Y(y: corners[index]), z: CGFloat(node.volume.coordinate.z) + unit.vertices[index].z))
+            vertices.append(SCNVector3(x: CGFloat(node.volume.coordinate.x) + CGFloat(unit.vertices[index].x), y: World.Y(y: corners[index]), z: CGFloat(node.volume.coordinate.z) + unit.vertices[index].z))
         }
         
         let upperPolytope = Polytope(vertices: vertices)
